# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MyBusiness 2.0 API",
        "description": "MyBusiness API documentation",
        "contact": {
            "email": "sgg.2020@gmail.com"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/college/index": {
            "get": {
                "tags": [
                    "College"
                ],
                "summary": "Get all college",
                "operationId": "indexCollege",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/college/status/{id}": {
            "get": {
                "tags": [
                    "College"
                ],
                "summary": "Get colleges by status",
                "operationId": "showCollegeByStatus",
                "parameters": [
                    {
                        "name": "status",
                        "in": "path",
                        "description": "Boolean ID",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "default": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/college/{id}": {
            "get": {
                "tags": [
                    "College"
                ],
                "summary": "Get college by ID",
                "operationId": "showCollege",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "String ID",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "D9B7E73F-9E5B-4884-AF72-1FE5934A6E3A"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/college/add": {
            "post": {
                "tags": [
                    "College"
                ],
                "summary": "Create college",
                "operationId": "ff97cbf68a594153206e73edf7288f6d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "code": {
                                                "type": "string"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "college_id": {
                                                "type": "uuid"
                                            },
                                            "active": {
                                                "type": "boolean"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "code": "MSG",
                                    "name": "Master",
                                    "description": "Degree details",
                                    "college_id": "D9B7E73F-9E5B-4884-AF72-1FE5934A6E3A",
                                    "active": "1"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "type": "object"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/college/update/{id}": {
            "put": {
                "tags": [
                    "College"
                ],
                "summary": "Update college",
                "operationId": "7bd11fb122b5162437ce113a063e1463",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Uuid",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "uuid",
                            "default": "D9B7E73F-9E5B-4884-AF72-1FE5934A6E3A"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "code": {
                                                "type": "string"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "college_id": {
                                                "type": "uuid"
                                            },
                                            "active": {
                                                "type": "boolean"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "code": "MSG",
                                    "name": "Master",
                                    "description": "Degree details",
                                    "college_id": "D9B7E73F-9E5B-4884-AF72-1FE5934A6E3A",
                                    "active": "1"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "type": "object"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/college/delete/{id}": {
            "delete": {
                "tags": [
                    "College"
                ],
                "summary": "Delete college by id",
                "operationId": "deleteCollege",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Uuid",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "0C6A36BA-10E4-438F-BA86-0D5B68A2BB15"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    }
}